billing_service/requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
kafka-python>=2.0.2
sqlalchemy>=1.4.47,<2.0.0 # Pin V1.4 for now, V2 requires syntax changes
psycopg2-binary>=2.9.3
pydantic>=1.10.7,<2.0.0 # Match toll_processor for BaseSettings V1 compatibility
pydantic-settings>=2.0.0
python-json-logger>=2.0.7
python-dotenv>=1.0.0
alembic>=1.10.3 # For database migrations
# httpx>=0.24.0 # If calling real external payment APIs asynchronously
